requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
huggingface_hub
streamlit
requests
pypdf
python-docx